requirements.txt
click
logzero
pygithub
pylint
